// src/app/components/types.ts
export interface Lead {
  NextFollowUpDate?: string;
  lastSync?: Date | string;
  id?: string;
  'Nama Perusahaan': string;
  PIC: string;
  Telepon: string;
  Layanan: string;
  Tahapan: string;
  Bulan: string;
  'Nilai Deal': number;
  Email?: string;
  Catatan?: string;
  'Last Updated'?: string;
  chatHistory?: string;
  namaPerusahaan?: string;
  pic?: string;
  telepon?: string;
  layanan?: string;
  tahapan?: string;
  bulan?: string;
  nilaiDeal?: number;
  email?: string;
  catatan?: string;
  updateTerakhir?: string;
}